Updated to make parameters editable
The donation amounts are in an array now so you can edit these to include any choices you'd like to offer your candidates.
The states are in an array now and include the US Territories.  If you don't need them, you can delete those entries.
The legal disclaimer message is now a variable at the top of the index file so you can change the language easily as needed.
Also made the page title and the page header a variable.
Extracted the CSS to be in an external file to make it easier to edit.